requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
crewai>=1.0.0
streamlit>=1.40.0
